+++ src/components/visual-query-builder/VisualEditor.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { EditorField, EditorRow, EditorRows } from '@grafana/experimental';
 import { QueryEditorProps, QueryRowFilter } from 'types';
 import { BQOrderByRow } from './BQOrderByRow';
 import { BQSelectRow } from './BQSelectRow';
@@ -7,6 +6,7 @@ import { BQWhereRow } from './BQWhereRow';
 import { Preview } from './Preview';
 import { BQGroupByRow } from './BQGroupByRow';
 import { QueryToolbox } from 'components/query-editor-raw/QueryToolbox';
+import { EditorRows, EditorRow, EditorField } from '@grafana/experimental';
 
 interface VisualEditorProps extends QueryEditorProps {
   queryRowFilter: QueryRowFilter;